Move Note and related classes into its own folder
