+++ S2VX.Game/Editor/Reversible/ReversibleUpdateNoteCoordinates.cs
@@ -1,5 +1,5 @@
 ﻿using osuTK;
-using S2VX.Game.Story;
+using S2VX.Game.Story.Note;
 
 namespace S2VX.Game.Editor.Reversible {
     public class ReversibleUpdateNoteCoordinates : IReversible {
