+++ S2VX.Game/Story/Note/Approach.cs
@@ -4,7 +4,7 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Utils;
 using osuTK;
 
-namespace S2VX.Game.Story {
+namespace S2VX.Game.Story.Note {
     public class Approach : CompositeDrawable {
         public double EndTime { get; set; }
         public Vector2 Coordinates { get; set; } = Vector2.Zero;
