+++ S2VX.Game/Story/Note/Approaches.cs
@@ -3,7 +3,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using System.Collections.Generic;
 
-namespace S2VX.Game.Story {
+namespace S2VX.Game.Story.Note {
     public class Approaches : CompositeDrawable {
         public List<Approach> Children { get; private set; } = new List<Approach>();
         public void SetChildren(List<Approach> approaches) => Children = approaches;
